+++ src/include/86box/snd_sb.h
@@ -222,8 +222,8 @@ extern void    sb_ess_mixer_write(uint16_t addr, uint8_t val, void *priv);
 extern uint8_t sb_ess_mixer_read(uint16_t addr, void *priv);
 extern void    sb_ess_mixer_reset(sb_t *sb);
 
-extern void sb_get_buffer_sbpro(int32_t *buffer, int len, void *priv);
-extern void sb_get_music_buffer_sbpro(int32_t *buffer, int len, void *priv);
+extern void sb_get_buffer_sbpro(int32_t *buffer, uint16_t len, void *priv);
+extern void sb_get_music_buffer_sbpro(int32_t *buffer, uint16_t len, void *priv);
 extern void sbpro_filter_cd_audio(int channel, double *buffer, void *priv);
 extern void sb16_awe32_filter_cd_audio(int channel, double *buffer, void *priv);
 extern void sb_close(void *priv);
